pyproject.toml
[tool.poetry]
name = "rag-law"
version = "0.1.0"
description = "Initial package list and dependencies"
authors = ["B8ni"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandas = "^2.2.3"
requests = "^2.32.3"
farm-haystack = "^1.26.4"
setuptools = "^75.8.0"
faiss-cpu = "^1.9.0.post1"
nltk = "^3.9.1"
pypdf2 = "^3.0.1"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
